requirements.txt
eyed3
discord.py
requests